gitbak

+++ bench/subsetsum/rho.cpp
@@ -13,7 +13,7 @@
 
 int main() {
 	constexpr uint32_t n = PARAM_n;
-	constexpr uint32_t q = 1ul << n;
+	constexpr uint64_t q = 1ul << n;
 	constexpr static SSS instance{.n=n, .q=q, .bp=PARAM_n1_3};
 	using S = sss_d2<instance>;
